# services/ai-gateway/openapi.yaml
openapi: 3.0.0
info:
  title: InsightPulse FinServ API
  description: AI-powered access to FinServ operations (Odoo ERP, Close Ops, Policy QA)
  version: 1.0.0
  contact:
    email: support@insightpulseai.net

servers:
  - url: https://api.insightpulseai.net
    description: Production API

security:
  - ApiKeyAuth: []

paths:
  # ============================================================================
  # Close Operations
  # ============================================================================

  /api/close/status:
    get:
      summary: Get month-end close status
      description: Returns close status for all entities or a specific entity
      operationId: getCloseStatus
      tags: [Close Operations]
      parameters:
        - name: entity_code
          in: query
          schema:
            type: string
            enum: [RIM, CKVC, BOM, JPAL, JLI, JAP, LAS, RMQB, CONSOLIDATED]
          description: Filter by entity code
        - name: period
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Period in YYYY-MM format
      responses:
        '200':
          description: Close status returned
          content:
            application/json:
              schema:
                type: object

  /api/close/checklist:
    get:
      summary: Get close checklist
      description: Returns detailed close checklist with task status
      operationId: getCloseChecklist
      tags: [Close Operations]
      parameters:
        - name: entity_code
          in: query
          required: true
          schema:
            type: string
            enum: [RIM, CKVC, BOM, JPAL, JLI, JAP, LAS, RMQB]
        - name: period
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Checklist returned

  /api/close/exceptions:
    get:
      summary: Get close exceptions
      description: Returns open exceptions/blockers for close
      operationId: getCloseExceptions
      tags: [Close Operations]
      parameters:
        - name: entity_code
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
      responses:
        '200':
          description: Exceptions returned

  /api/close/metrics:
    get:
      summary: Get close metrics
      description: Returns close cycle metrics (cycle time, exceptions, trends)
      operationId: getCloseMetrics
      tags: [Close Operations]
      parameters:
        - name: entity_code
          in: query
          schema:
            type: string
        - name: last_n_periods
          in: query
          schema:
            type: integer
            default: 6
      responses:
        '200':
          description: Metrics returned

  # ============================================================================
  # Policy QA
  # ============================================================================

  /api/policy/qa:
    post:
      summary: Ask a policy question
      description: Ask a question about company policies, SOPs, or compliance. Returns answer with citations.
      operationId: policyQA
      tags: [Policy & Compliance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  description: Your question about policies/procedures
                  example: "What is the SLA for month-end close accruals?"
                category:
                  type: string
                  enum: [accounting, finance, hr, procurement, compliance, bir, all]
                  default: all
                top_k:
                  type: integer
                  default: 6
      responses:
        '200':
          description: Answer with citations returned

  /api/policy/search:
    post:
      summary: Search policy documents
      description: Search for policies by keyword or phrase
      operationId: policySearch
      tags: [Policy & Compliance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  example: "expense approval workflow"
                category:
                  type: string
                  enum: [accounting, finance, hr, procurement, compliance, bir, all]
                  default: all
                limit:
                  type: integer
                  default: 10
      responses:
        '200':
          description: Search results returned

  /api/policy/categories:
    get:
      summary: List policy categories
      description: Returns all available policy categories with document counts
      operationId: listPolicyCategories
      tags: [Policy & Compliance]
      responses:
        '200':
          description: Categories returned

  # ============================================================================
  # Odoo ERP Operations
  # ============================================================================

  /api/expenses:
    get:
      summary: List expenses
      description: Get list of expense reports with optional filters
      operationId: listExpenses
      tags: [Expenses]
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
            enum: [draft, reported, approved, done, refused]
        - name: agency_code
          in: query
          schema:
            type: string
            enum: [RIM, CKVC, BOM, JPAL, JLI, JAP, LAS, RMQB]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Expenses returned

  /api/expenses/{expense_ref}:
    get:
      summary: Get expense details
      description: Get detailed information about a specific expense
      operationId: getExpense
      tags: [Expenses]
      parameters:
        - name: expense_ref
          in: path
          required: true
          schema:
            type: string
          description: Expense number (e.g., EXP-001) or ID
      responses:
        '200':
          description: Expense details returned

  /api/invoices:
    get:
      summary: List invoices
      description: Get list of invoices with filters
      operationId: listInvoices
      tags: [Invoices]
      parameters:
        - name: partner_id
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
            enum: [draft, posted, cancel]
        - name: invoice_type
          in: query
          schema:
            type: string
            enum: [out_invoice, in_invoice, out_refund, in_refund]
        - name: overdue
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Invoices returned

  /api/sale-orders/{order_ref}:
    get:
      summary: Get sale order
      description: Get detailed information about a sale order
      operationId: getSaleOrder
      tags: [Sales]
      parameters:
        - name: order_ref
          in: path
          required: true
          schema:
            type: string
          description: Sale order number (e.g., SO001) or ID
      responses:
        '200':
          description: Sale order details returned

  /api/tasks:
    post:
      summary: Create task
      description: Create a new project task in Odoo
      operationId: createTask
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Task title
                description:
                  type: string
                  description: Task description (markdown supported)
                project_id:
                  type: integer
                  description: Project ID
                assigned_to:
                  type: integer
                  description: User ID to assign to
                priority:
                  type: string
                  enum: ['0', '1', '2', '3']
                  default: '1'
                  description: Priority (0=Low, 1=Normal, 2=High, 3=Urgent)
      responses:
        '200':
          description: Task created

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

tags:
  - name: Close Operations
    description: Month-end close status, checklists, and metrics
  - name: Policy & Compliance
    description: Policy Q&A and document search with RAG
  - name: Expenses
    description: Expense report operations
  - name: Invoices
    description: Invoice operations
  - name: Sales
    description: Sales order operations
  - name: Tasks
    description: Project task operations